+++ tests/index.js
@@ -1,8 +1,12 @@
 window.onload = function() {
   // There's options in this code
-  var typing = new Typing("#typing", {
-    color: "orange",
-    speed: 40,
-    transition: 150
-  });
+  var options = {
+    color: "black",
+    speed: 60,
+    transition: 150, 
+    width: 1.5,
+    string: "Typing.js is working successfuly"
+  }
+
+  var typing = new Typing("#typing", options);
 }