+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Coming soon for the future versions...
 
+## 1.0.3
+- Added promise to callback private functions
+- Added `string` variable in the options
+- Code fixed
+
+## 1.0.2
+- Adding docs and demos and publishing it in Github Pages (failed)
+
 ## 1.0.1
 - Code fixed
 - Functions added `on()` and `off()`